moved Tristate and operator !(Tristate) to Tristate.enh.h
